Optimize encoding for cells
Naively saving the whole thing as JSON ended up weighing in at 40MB,
which is _WAY_ over localStorage quota.

Turns out we already have a really great, compact way to store text:
text. Even better, we already have serializing/deseriaizing functions we
can use: outputText() and fromText(). This brings the size down to 1.2MB
in the worst case where all 2000x600 cells are in use; in practice, the
size is tiny because only the cells that are in use are saved via
outputText().